Modified detection of wires crossing
Defined a new function to detect wires crossing like a "T". Added a comment in cpdiodes.js to a problem when trying to change a normal wire connection as a "T" to a diode connection. No idea for a quick fix as of this commit. I'm sure that my idea (explained in the comment) will break many things because of the project structure.