First try create chain of responsibility
Created a file with first building blocks to implement the wire connection change between normal/no/diode connection as a chain of responsibility. I didn't see any possible quick fix for the problem found in the previous commit. Not included in main code yet and that might be tricky. Might need to replace the conpoints array with a Map to keep track of the connection type.