Partially fixed T wire connection transitions
T wire connections now switch directly between normal <-> diode connections. Still when trying to change from full cross connection with diode to a T connection, it resets to a normal dot connection., Problem is in deleteWires(). Not sure whether changing the code breaks anything unexpected, so leave it for now.